refactor(reposicao): extrai 8 render helpers puros de AdminRoutePlanner
Segundo corte do RoutePlanner. Move pra routePlanner/renderHelpers.tsx
as 8 funções de apresentação puras (nenhuma fecha sobre estado do
componente): formatTimer, formatDuration, openInWaze, openInGoogleMaps,
getStopIcon, getCTALabel, getVisitBadge, getOrderBadge.

Verbatim — zero mudança de comportamento. Inclui preservar um bug
pré-existente conhecido no formatDuration (`min % h` deveria ser
`min % 60`), rastreado em task separada (não corrigido aqui pra manter
o corte como puro move).

renderHelpers.tsx entra no strict (347). Nenhum import lucide órfão
criado (Truck/ShoppingBag/Layers/Users seguem usados no corpo).

Verificado: tsc 0, typecheck:strict 0, eslint-unused 0, build OK, 392
testes passando.
