+++ src/pages/AdminRoutePlanner.tsx
@@ -32,6 +32,16 @@ import {
   PRIORITY_CONFIG,
   STOP_CONFIG,
 } from '@/components/reposicao/routePlanner/constants';
+import {
+  formatTimer,
+  formatDuration,
+  openInWaze,
+  openInGoogleMaps,
+  getStopIcon,
+  getCTALabel,
+  getVisitBadge,
+  getOrderBadge,
+} from '@/components/reposicao/routePlanner/renderHelpers';
 
 // Fix default marker icons for Leaflet + bundlers
 delete (L.Icon.Default.prototype as any)._getIconUrl;
@@ -471,11 +481,6 @@ const AdminRoutePlanner = () => {
     }
   };
   
-  const formatTimer = (seconds: number) => {
-    const m = Math.floor(seconds / 60);
-    const s = seconds % 60;
-    return `${m}:${String(s).padStart(2, '0')}`;
-  };
 
   const loadCommercialStops = async () => {
     try {
@@ -843,60 +848,6 @@ const AdminRoutePlanner = () => {
     leafletMap.current.fitBounds(bounds, { padding: [40, 40] });
   }, [optimizedRoute]);
 
-  const openInWaze = (stop: RouteStop) => {
-    if (stop.lat && stop.lng) {
-      window.open(`https://waze.com/ul?ll=${stop.lat},${stop.lng}&navigate=yes`, '_blank');
-    } else {
-      const q = `${stop.address.street}, ${stop.address.number}, ${stop.address.city}, ${stop.address.state}`;
-      window.open(`https://waze.com/ul?q=${encodeURIComponent(q)}&navigate=yes`, '_blank');
-    }
-  };
-
-  const openInGoogleMaps = (stop: RouteStop) => {
-    if (stop.lat && stop.lng) {
-      window.open(`https://www.google.com/maps/dir/?api=1&destination=${stop.lat},${stop.lng}`, '_blank');
-    } else {
-      const q = `${stop.address.street}, ${stop.address.number}, ${stop.address.city}, ${stop.address.state}`;
-      window.open(`https://www.google.com/maps/dir/?api=1&destination=${encodeURIComponent(q)}`, '_blank');
-    }
-  };
-
-  const getStopIcon = (type: StopType) => {
-    switch (type) {
-      case 'pickup_tools': return <Truck className="w-3.5 h-3.5" />;
-      case 'deliver_tools': return <Truck className="w-3.5 h-3.5" />;
-      case 'sales_visit': return <ShoppingBag className="w-3.5 h-3.5" />;
-      case 'hybrid_visit': return <Layers className="w-3.5 h-3.5" />;
-      case 'manual_visit': return <Users className="w-3.5 h-3.5" />;
-    }
-  };
-  
-  const getVisitBadge = (customer: ManualCustomer) => {
-    if (customer.daysSinceLastVisit === null) {
-      return <Badge variant="danger" className="text-xs">Nunca visitado</Badge>;
-    }
-    if (customer.daysSinceLastVisit > 30) {
-      return <Badge variant="warning" className="text-xs">Última visita há {customer.daysSinceLastVisit} dias</Badge>;
-    }
-    return null;
-  };
-  
-  const getOrderBadge = (customer: ManualCustomer) => {
-    if (customer.daysSinceLastOrder === null) {
-      return null;
-    }
-    if (customer.daysSinceLastOrder > 90) {
-      return <Badge variant="danger" className="text-xs">Sem compra há {customer.daysSinceLastOrder} dias</Badge>;
-    }
-    if (customer.daysSinceLastOrder > 30) {
-      return <Badge variant="warning" className="text-xs">Comprou há {customer.daysSinceLastOrder} dias</Badge>;
-    }
-    if (customer.daysSinceLastOrder <= 30) {
-      return <Badge variant="success" className="text-xs">Comprou recentemente</Badge>;
-    }
-    return null;
-  };
-  
   const filteredManualCustomers = useMemo(() => {
     let filtered = manualCustomers;
     
@@ -946,12 +897,6 @@ const AdminRoutePlanner = () => {
     }
   };
 
-  const getCTALabel = (stop: RouteStop) => {
-    if (stop.orderId) return 'Ver pedido';
-    if (stop.visitReason.includes('afiação vencida')) return 'Criar pedido de afiação';
-    return 'Criar pedido';
-  };
-
   // Stats
   const stopCounts = useMemo(() => {
     const counts = { pickup_tools: 0, deliver_tools: 0, sales_visit: 0, hybrid_visit: 0 };
@@ -978,13 +923,6 @@ const AdminRoutePlanner = () => {
     }
     return { stopMin, travelMin, totalMin: stopMin + travelMin };
   }, [optimizedRoute]);
-
-  const formatDuration = (min: number) => {
-    if (min < 60) return `${min}min`;
-    const h = Math.floor(min / 60);
-    const m = min % h;
-    return m > 0 ? `${h}h${String(m).padStart(2, '0')}` : `${h}h`;
-  };
 
   const isLoading = authLoading || loading;
